+++ BirdWatcherBackend/Models/devTempLight.cs
@@ -5,8 +5,8 @@ namespace BirdWatcherBackend.Models
     public class devTempLight
     {
         public long Id { get; set; }
-        public int Temperature { get; set; }
-        public int PhotoresistorValue { get; set; }
+        public float Temperature { get; set; }
+        public float PhotoresistorValue { get; set; }
         public DateTime Timestamp { get; set; }
     }
 }
